requirements.txt
alignment==1.0.10
captum==0.8.0
matplotlib==3.10.3
numpy==2.3.1
pandas==2.3.0
Pillow==11.2.1
scikit_learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
statsmodels==0.14.4
timm==1.0.15
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
